Fold build assertions into the serve.dynamic-routing test case
