+++ packages/cli/test/cases/serve.dynamic-routing/src/pages/[json].js
